resolve Conflict
